+++ src/sync.ts
@@ -18,6 +18,8 @@ import {
 
 import PragmaUtil from "./pragmaUtil";
 
+let globalCommonService: Commons;
+
 export class Sync {
   constructor(private context: vscode.ExtensionContext) {}
   /**
@@ -25,8 +27,7 @@ export class Sync {
    */
   public async bootstrap(): Promise<void> {
     const env = new Environment(this.context);
-    const globalCommonService = new Commons(env, this.context);
-
+    globalCommonService = new Commons(env, this.context);
     // if lock file not exist
     // then create it
     if (!(await FileService.FileExists(env.FILE_SYNC_LOCK))) {
@@ -53,17 +54,24 @@ export class Sync {
         return;
       }
 
-      if (gistAvailable === true && startUpSetting.autoDownload === true) {
-        vscode.commands
-          .executeCommand("extension.downloadSettings")
-          .then(() => {
-            if (startUpSetting.autoUpload && tokenAvailable && gistAvailable) {
-              return globalCommonService.StartWatch();
-            }
-          });
-      }
-      if (startUpSetting.autoUpload && tokenAvailable && gistAvailable) {
-        return globalCommonService.StartWatch();
+      if (gistAvailable) {
+        if (startUpSetting.autoDownload) {
+          vscode.commands
+            .executeCommand("extension.downloadSettings")
+            .then(() => {
+              if (
+                startUpSetting.autoUpload &&
+                tokenAvailable &&
+                gistAvailable
+              ) {
+                return globalCommonService.autoUploadService.StartWatching();
+              }
+            });
+        } else {
+          if (startUpSetting.autoUpload && tokenAvailable && gistAvailable) {
+            return globalCommonService.autoUploadService.StartWatching();
+          }
+        }
       }
     }
   }
@@ -71,19 +79,19 @@ export class Sync {
    * Upload setting to github gist
    */
   public async upload(): Promise<void> {
+    // @ts-ignore
     const args = arguments;
     const env = new Environment(this.context);
-    const common = new Commons(env, this.context);
     let github: GitHubService = null;
     let localConfig: LocalConfig = new LocalConfig();
     const allSettingFiles: File[] = [];
     let uploadedExtensions: ExtensionInformation[] = [];
     const ignoredExtensions: ExtensionInformation[] = [];
     const dateNow = new Date();
-    common.CloseWatch();
+    globalCommonService.autoUploadService.StopWatching();
 
     try {
-      localConfig = await common.InitalizeSettings();
+      localConfig = await globalCommonService.InitalizeSettings();
       localConfig.publicGist = false;
       if (args.length > 0) {
         if (args[0] === "publicGIST") {
@@ -95,11 +103,9 @@ export class Sync {
         localConfig.customConfig.token,
         localConfig.customConfig.githubEnterpriseUrl
       );
-      // ignoreSettings = await common.GetIgnoredSettings(localConfig.customConfig.ignoreUploadSettings);
       await startGitProcess(localConfig.extConfig, localConfig.customConfig);
-      // await common.SetIgnoredSettings(ignoreSettings);
     } catch (error) {
-      Commons.LogException(error, common.ERROR_MESSAGE, true);
+      Commons.LogException(error, globalCommonService.ERROR_MESSAGE, true);
       return;
     }
 
@@ -190,10 +196,10 @@ export class Sync {
           contentFiles = contentFiles.filter((contentFile: File) => {
             const matchedFolders = customSettings.ignoreUploadFolders.filter(
               folder => {
-                return contentFile.filePath.indexOf(folder) === -1;
+                return contentFile.filePath.indexOf(folder) !== -1;
               }
             );
-            return matchedFolders.length > 0;
+            return matchedFolders.length === 0;
           });
         }
         const customFileKeys: string[] = Object.keys(
@@ -209,7 +215,7 @@ export class Sync {
           }
         }
       } else {
-        Commons.LogException(null, common.ERROR_MESSAGE, true);
+        Commons.LogException(null, globalCommonService.ERROR_MESSAGE, true);
         return;
       }
 
@@ -251,7 +257,7 @@ export class Sync {
       try {
         if (syncSetting.gist == null || syncSetting.gist === "") {
           if (customSettings.askGistName) {
-            customSettings.gistDescription = await common.AskGistName();
+            customSettings.gistDescription = await globalCommonService.AskGistName();
           }
           newGIST = true;
           const gistID = await github.CreateEmptyGIST(
@@ -318,14 +324,16 @@ export class Sync {
           return;
         }
       } catch (err) {
-        Commons.LogException(err, common.ERROR_MESSAGE, true);
+        Commons.LogException(err, globalCommonService.ERROR_MESSAGE, true);
         return;
       }
 
       if (completed) {
         try {
-          const settingsUpdated = await common.SaveSettings(syncSetting);
-          const customSettingsUpdated = await common.SetCustomSettings(
+          const settingsUpdated = await globalCommonService.SaveSettings(
+            syncSetting
+          );
+          const customSettingsUpdated = await globalCommonService.SetCustomSettings(
             customSettings
           );
           if (settingsUpdated && customSettingsUpdated) {
@@ -345,7 +353,7 @@ export class Sync {
             }
 
             if (!syncSetting.quietSync) {
-              common.ShowSummaryOutput(
+              globalCommonService.ShowSummaryOutput(
                 true,
                 allSettingFiles,
                 null,
@@ -362,11 +370,11 @@ export class Sync {
               );
             }
             if (syncSetting.autoUpload) {
-              common.StartWatch();
+              globalCommonService.autoUploadService.StartWatching();
             }
           }
         } catch (err) {
-          Commons.LogException(err, common.ERROR_MESSAGE, true);
+          Commons.LogException(err, globalCommonService.ERROR_MESSAGE, true);
         }
       }
     }
@@ -376,15 +384,14 @@ export class Sync {
    */
   public async download(): Promise<void> {
     const env = new Environment(this.context);
-    const common = new Commons(env, this.context);
     let localSettings: LocalConfig = new LocalConfig();
-    common.CloseWatch();
+    globalCommonService.autoUploadService.StopWatching();
 
     try {
-      localSettings = await common.InitalizeSettings();
+      localSettings = await globalCommonService.InitalizeSettings();
       await StartDownload(localSettings.extConfig, localSettings.customConfig);
     } catch (err) {
-      Commons.LogException(err, common.ERROR_MESSAGE, true);
+      Commons.LogException(err, globalCommonService.ERROR_MESSAGE, true);
       return;
     }
 
@@ -551,8 +558,6 @@ export class Sync {
 
                 addedExtensions = await PluginService.InstallExtensions(
                   content,
-                  env.ExtensionFolder,
-                  useCli,
                   ignoredExtensions,
                   env.OsType,
                   env.isInsiders,
@@ -618,7 +623,11 @@ export class Sync {
                     // TODO : add Name attribute in File and show information message here with name , when required.
                   })
                   .catch(err => {
-                    Commons.LogException(err, common.ERROR_MESSAGE, true);
+                    Commons.LogException(
+                      err,
+                      globalCommonService.ERROR_MESSAGE,
+                      true
+                    );
                     return;
                   })
               );
@@ -628,13 +637,15 @@ export class Sync {
       }
 
       await Promise.all(actionList);
-      const settingsUpdated = await common.SaveSettings(syncSetting);
-      const customSettingsUpdated = await common.SetCustomSettings(
+      const settingsUpdated = await globalCommonService.SaveSettings(
+        syncSetting
+      );
+      const customSettingsUpdated = await globalCommonService.SetCustomSettings(
         customSettings
       );
       if (settingsUpdated && customSettingsUpdated) {
         if (!syncSetting.quietSync) {
-          common.ShowSummaryOutput(
+          globalCommonService.ShowSummaryOutput(
             false,
             updatedFiles,
             deletedExtensions,
@@ -658,19 +669,8 @@ export class Sync {
             5000
           );
         }
-        if (Object.keys(customSettings.replaceCodeSettings).length > 0) {
-          const config = vscode.workspace.getConfiguration();
-          const keysDefined: string[] = Object.keys(
-            customSettings.replaceCodeSettings
-          );
-          for (const key of keysDefined) {
-            const value: string = customSettings.replaceCodeSettings[key];
-            const c: any = value === "" ? undefined : value;
-            config.update(key, c, true);
-          }
-        }
         if (syncSetting.autoUpload) {
-          common.StartWatch();
+          globalCommonService.autoUploadService.StartWatching();
         }
       } else {
         vscode.window.showErrorMessage(
@@ -767,7 +767,6 @@ export class Sync {
       "cmd.otherOptions.toggleAutoUpload",
       "cmd.otherOptions.toggleAutoDownload",
       "cmd.otherOptions.toggleSummaryPage",
-      "cmd.otherOptions.preserve",
       "cmd.otherOptions.customizedSync",
       "cmd.otherOptions.downloadCustomFile",
       "cmd.otherOptions.joinCommunity",
@@ -789,9 +788,6 @@ export class Sync {
 
     const handlerMap = {
       0: async () => {
-        common.OpenSettingsPage();
-      },
-      1: async () => {
         const file: vscode.Uri = vscode.Uri.file(env.FILE_CUSTOMIZEDSETTINGS);
         fs.openSync(file.fsPath, "r");
         const document = await vscode.workspace.openTextDocument(file);
@@ -801,7 +797,7 @@ export class Sync {
           true
         );
       },
-      2: async () => {
+      1: async () => {
         // share public gist
         const answer = await vscode.window.showInformationMessage(
           localize("cmd.otherOptions.shareSetting.beforeConfirm"),
@@ -817,26 +813,26 @@ export class Sync {
           await common.SetCustomSettings(customSettings);
         }
       },
-      3: async () => {
+      2: async () => {
         // Download Settings from Public GIST
         selectedItem = 2;
         customSettings.downloadPublicGist = true;
         settingChanged = true;
         await common.SetCustomSettings(customSettings);
       },
-      4: async () => {
+      3: async () => {
         // toggle force download
         selectedItem = 3;
         settingChanged = true;
         setting.forceDownload = !setting.forceDownload;
       },
-      5: async () => {
+      4: async () => {
         // toggle auto upload
         selectedItem = 4;
         settingChanged = true;
         setting.autoUpload = !setting.autoUpload;
       },
-      6: async () => {
+      5: async () => {
         // auto download on startup
         selectedItem = 5;
         settingChanged = true;
@@ -851,7 +847,7 @@ export class Sync {
 
         setting.autoDownload = !setting.autoDownload;
       },
-      7: async () => {
+      6: async () => {
         // page summary toggle
         selectedItem = 6;
         settingChanged = true;
@@ -862,7 +858,7 @@ export class Sync {
         }
         setting.quietSync = !setting.quietSync;
       },
-      8: async () => {
+      7: async () => {
         // preserve
         const options: vscode.InputBoxOptions = {
           ignoreFocusOut: true,
@@ -890,7 +886,7 @@ export class Sync {
           }
         }
       },
-      9: async () => {
+      8: async () => {
         // add customized sync file
         const options: vscode.InputBoxOptions = {
           ignoreFocusOut: true,
@@ -913,7 +909,7 @@ export class Sync {
           }
         }
       },
-      10: async () => {
+      9: async () => {
         // Import customized sync file to workspace
         const customFiles = await this.getCustomFilesFromGist(
           customSettings,
@@ -957,7 +953,7 @@ export class Sync {
           }
         }
       },
-      11: async () => {
+      10: async () => {
         vscode.commands.executeCommand(
           "vscode.open",
           vscode.Uri.parse(
@@ -965,7 +961,7 @@ export class Sync {
           )
         );
       },
-      12: async () => {
+      11: async () => {
         vscode.commands.executeCommand(
           "vscode.open",
           vscode.Uri.parse(
@@ -973,7 +969,7 @@ export class Sync {
           )
         );
       },
-      13: async () => {
+      12: async () => {
         vscode.commands.executeCommand(
           "vscode.open",
           vscode.Uri.parse(
@@ -987,7 +983,7 @@ export class Sync {
       await handlerMap[index]();
       if (settingChanged) {
         if (selectedItem === 1) {
-          common.CloseWatch();
+          globalCommonService.autoUploadService.StopWatching();
         }
         await common
           .SaveSettings(setting)